docs: updated documentation and removed unused files
