--- ios/DynamicWidgetExtensionAttributes.swift
@@ -1,23 +1,0 @@
-import ActivityKit
-
-struct DynamicWidgetExtensionAttributes: ActivityAttributes {
-    public struct ContentState: Codable, Hashable {
-        var bodyComponentId: String
-        var lockScreenComponentId: String
-        var compactLeadingComponentId: String
-        var compactTrailingComponentId: String
-        var minimalComponentId: String
-        
-        init(lockScreenComponentId: String = "",
-             bodyComponentId: String = "",
-             compactLeadingComponentId: String = "",
-             compactTrailingComponentId: String = "",
-             minimalComponentId: String = "") {
-            self.lockScreenComponentId = lockScreenComponentId
-            self.bodyComponentId = bodyComponentId
-            self.compactLeadingComponentId = compactLeadingComponentId
-            self.compactTrailingComponentId = compactTrailingComponentId
-            self.minimalComponentId = minimalComponentId
-        }
-    }
-}
